// src/utils/common.ts
import { IMonitorDocument } from "@app/interfaces/monitor.interface";
import { getAllUsersActiveMonitors, getMonitorById, getUserActiveMonitors, startCreatedMonitors } from "@app/services/monitor.service";
import { toLower } from "lodash";
import { setTimeout } from "timers";
import { appTimeZone } from "./index"
import { startSingleJob } from "./job"
import { verify } from 'jsonwebtoken';
import { config } from "../server/config"
import { IAuthPayload } from "../interfaces/user.interface"
import { pubSub } from "./index"
import { IEmailLocals } from "@app/interfaces/notification.interface";
import { IHeartbeat } from "@app/interfaces/heartbeat.interface";
import { sendEmail } from "../server/email";
import { ISSLMonitorDocument } from "../interfaces/ssl.interface"
import { getAllUsersActiveSSLMonitors, sslStatusMonitor, getSSLMonitorById } from "../services/ssl.service"

export const generateRandomId = () => {
    return Math.floor(Math.random() * (9999 - 1000 + 1)) + 1000;
}

export const getRandomInt = (min: number, max: number): number => {
    min = Math.ceil(min);
    max = Math.floor(max);
    return Math.floor(Math.random() * (max - min + 1)) + min;
};


export const sleep = (ms: number): Promise<void> => {
    return new Promise((resolve) => setTimeout(resolve, ms));
};

export const startMonitors = async (): Promise<void> => {
    const list: IMonitorDocument[] = await getAllUsersActiveMonitors();
    console.log("Monitor list ------->", list);
    for (const monitor of list) {
        console.log("oooooooooooooooo");
        startCreatedMonitors(monitor, toLower(monitor.name), monitor.type);
        await sleep(getRandomInt(300, 1000));
    }
};

export const resumeMonitors = async (monitorId: number): Promise<void> => {
    console.log("resume");
    const monitor: IMonitorDocument = await getMonitorById(monitorId);
    startCreatedMonitors(monitor, toLower(monitor.name), monitor.type);
    await sleep(getRandomInt(300, 1000));
};

export const enableAutoRefreshJob = (cookies: string): void => {
    const result: Record<string, string> = getCookies(cookies);
    const session: string = Buffer.from(result.session, 'base64').toString();
    const payload: IAuthPayload = verify(JSON.parse(session).jwt, config.JWT_TOKEN) as IAuthPayload;
    const enableAutoRefresh: boolean = JSON.parse(session).enableAutomaticRefresh;
    console.log("---->", enableAutoRefresh);
    if (enableAutoRefresh) {
        startSingleJob(`${toLower(payload.username)}`, appTimeZone, 10, async () => {
            const monitors: IMonitorDocument[] = await getUserActiveMonitors(payload.id);
            pubSub.publish('MONITORS_UPDATED', {
                monitorsUpdated: {
                    userId: payload.id,
                    monitors
                }
            });
        });
    }
};

const getCookies = (cookie: string): Record<string, string> => {
    const cookies: Record<string, string> = {};
    cookie.split(';').forEach((cookieData) => {
        const parts: RegExpMatchArray | null = cookieData.match(/(.*?)=(.*)$/);
        cookies[parts![1].trim()] = (parts![2] || '').trim();
    });
    return cookies;
};

export const locals = (): IEmailLocals => {
    return {
        appLink: `${config.CLIENT_URL}`,
        appIcon: 'https://ibb.com/jD45fqX',
        appName: ''
    };
};

export const emailSender = async (notificationEmails: string, template: string, locals: IEmailLocals): Promise<void> => {
    console.log(notificationEmails);
    const emails = JSON.parse(notificationEmails);
    console.log("Parsed email --------------------->", emails);
    for (const email of emails) {
        await sendEmail(template, email, locals);
    }
};


export const encodeBase64 = (user: string, pass: string): string => {
    return Buffer.from(`${user || ''}:${pass || ''}`).toString('base64');
};

export const uptimePercentage = (heartbeats: IHeartbeat[]): number => {
    if (!heartbeats) {
        return 0;
    }
    const totalHeartbeats: number = heartbeats.length;
    const downtimeHeartbeats: number = heartbeats.filter((heartbeat: IHeartbeat) => heartbeat.status === 1).length;
    return Math.round(((totalHeartbeats - downtimeHeartbeats) / totalHeartbeats) * 100) || 0;
};

export const startSSLMonitors = async (): Promise<void> => {
    const list: ISSLMonitorDocument[] = await getAllUsersActiveSSLMonitors();

    for (const monitor of list) {
        sslStatusMonitor(monitor, toLower(monitor.name));
        await sleep(getRandomInt(300, 1000));
    }
};

export const resumeSSLMonitors = async (monitorId: number): Promise<void> => {
    const monitor: ISSLMonitorDocument = await getSSLMonitorById(monitorId);
    sslStatusMonitor(monitor, toLower(monitor.name));
    await sleep(getRandomInt(300, 1000));
};

export const getDaysBetween = (start: Date, end: Date): number => {
    return Math.round(Math.abs(+start - +end) / (1000 * 60 * 60 * 24));
};

export const getDaysRemaining = (start: Date, end: Date): number => {
    const daysRemaining = getDaysBetween(start, end);
    if (new Date(end).getTime() < new Date().getTime()) {
        return -daysRemaining;
    }
    return daysRemaining;
};
